:sparkles: feat(codemirror): expose loadCodeMirror loader via subpath
Move the CDN loader (`loadCodeMirror`, `ICodeMirror`,
`ICodeMirrorInstance`, `CodeMirrorOptions`) from
`src/plugins/codemirror-block/lib/index.ts` into `src/codemirror/loader.ts`
and re-export it through `@lobehub/editor/codemirror`. The plugin's `lib`
barrel now re-exports from the new location so existing internal imports
keep working unchanged. This lets downstream consumers actually
instantiate a CodeMirror editor (textarea + theme + Toolbar) without
pulling in the Lexical plugin layer.
