+++ src/codemirror/index.ts
@@ -3,6 +3,12 @@ export { LanguageSelect } from './components/LanguageSelect';
 export { MoreOptions } from './components/MoreOptions';
 export { Toolbar } from './components/Toolbar';
 export { LANGUAGES } from './constants';
+export {
+  type CodeMirrorOptions,
+  type ICodeMirror,
+  type ICodeMirrorInstance,
+  loadCodeMirror,
+} from './loader';
 export { styles } from './style';
 export { lobeTheme } from './theme';
 export type {
